requirements.txt
asgiref==3.8.1
Django==5.1.3
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.31
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
